Trainers: refactor tests
