+++ tests/trainers/test_utils.py
@@ -3,7 +3,7 @@
 
 import os
 from pathlib import Path
-from typing import Any, cast
+from typing import cast
 
 import pytest
 import timm
@@ -19,39 +19,8 @@ from torchgeo.trainers.utils import (
 )
 
 
-class ClassificationTestModel(Module):
-    def __init__(
-        self, in_chans: int = 3, num_classes: int = 1000, **kwargs: Any
-    ) -> None:
-        super().__init__()
-        self.conv1 = nn.Conv2d(in_channels=in_chans, out_channels=1, kernel_size=1)
-        self.pool = nn.AdaptiveAvgPool2d((1, 1))
-        self.fc = nn.Linear(1, num_classes)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        x = self.conv1(x)
-        x = self.pool(x)
-        x = torch.flatten(x, 1)
-        x = self.fc(x)
-        return x
-
-
-class RegressionTestModel(ClassificationTestModel):
-    def __init__(self, **kwargs: Any) -> None:
-        super().__init__(in_chans=3, num_classes=1)
-
-
-class SegmentationTestModel(Module):
-    def __init__(
-        self, in_channels: int = 3, classes: int = 1000, **kwargs: Any
-    ) -> None:
-        super().__init__()
-        self.conv1 = nn.Conv2d(
-            in_channels=in_channels, out_channels=classes, kernel_size=1, padding=0
-        )
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        return cast(torch.Tensor, self.conv1(x))
+def test_extract_backbone(checkpoint: str) -> None:
+    extract_backbone(checkpoint)
 
 
 def test_extract_backbone_unsupported_model(tmp_path: Path) -> None:
@@ -63,8 +32,11 @@ def test_extract_backbone_unsupported_model(tmp_path: Path) -> None:
         extract_backbone(path)
 
 
-def test_extract_backbone(checkpoint: str) -> None:
-    extract_backbone(checkpoint)
+def test_get_input_layer_name_and_module() -> None:
+    key, module = _get_input_layer_name_and_module(timm.create_model("resnet18"))
+    assert key == "conv1"
+    assert isinstance(module, nn.Conv2d)
+    assert module.in_channels == 3
 
 
 def test_load_state_dict(checkpoint: str, model: Module) -> None:
@@ -117,10 +89,3 @@ def test_reinit_initial_conv_layer() -> None:
     assert in_channels == 4
     assert k1 == 3 and k2 == 3
     assert new_conv_layer.stride[0] == 2
-
-
-def test_get_input_layer_name_and_module() -> None:
-    key, module = _get_input_layer_name_and_module(timm.create_model("resnet18"))
-    assert key == "conv1"
-    assert isinstance(module, nn.Conv2d)
-    assert module.in_channels == 3
